EDM-4833: test skip-vs-write via shouldPersistRenderedUpdate
